+++ src/Ext/Rules/Body.h
@@ -31,6 +31,7 @@ public:
 	public:
 		std::vector<std::vector<TechnoTypeClass*>> AITargetTypesLists;
 		std::vector<std::vector<ScriptTypeClass*>> AIScriptsLists;
+		std::vector<std::vector<HouseTypeClass*>> AIHousesLists;
 
 		Valueable<int> Storage_TiberiumIndex;
 		Valueable<float> HarvesterDumpAmount;