+++ src/Ext/TAction/Body.h
@@ -18,6 +18,7 @@ enum class PhobosTriggerAction : unsigned int
 	BinaryOperation = 504,
 	RunSuperWeaponAtLocation = 505,
 	RunSuperWeaponAtWaypoint = 506,
+	PrintMessageRemainingTechnos = 507,
 	ToggleMCVRedeploy = 510,
 
 	EditAngerNode = 606,
@@ -70,6 +71,7 @@ public:
 	ACTION_FUNC(BinaryOperation);
 	ACTION_FUNC(RunSuperWeaponAtLocation);
 	ACTION_FUNC(RunSuperWeaponAtWaypoint);
+	ACTION_FUNC(PrintMessageRemainingTechnos);
 	ACTION_FUNC(ToggleMCVRedeploy);
 
 	ACTION_FUNC(EditAngerNode);